\name{HW1Mantautas-package}
\alias{HW1Mantautas-package}
\alias{HW1Mantautas}
\docType{package}
\title{
\packageTitle{HW1Mantautas}
}
\description{
\packageDescription{HW1Mantautas}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HW1Mantautas}
\packageIndices{HW1Mantautas}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HW1Mantautas}

Maintainer: \packageMaintainer{HW1Mantautas}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
